feature: web delete task, isolated popup component
